Automatically build merge.lift and fix up the corner cases with sed
reposurgeon doesn't seem to take timestamp indices on the first element
of a set so we have to manually resolve those to the right mark.

Also if a merged branch is deleted, all commits back to the last tag are
lost, so only delete branches with git after we're done with rewriting.